requirements.txt
azure-functions
requests
azure-search-documents==11.4.0
